pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "me_250_790"
version = "0.1.0"
description = "ME.250.790 Utility code"
authors = [
    { name = "hannahburkhardt" }
]
dependencies = [
    "python-dotenv>=0.19.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "altair>=4.2.0",
    "altair_pandas @ git+https://github.com/altair-viz/altair_pandas",
    "seaborn>=0.11.0",
    "kagglehub",
    "matplotlib",
    "vegafusion",
    "vl-convert-python",
    "tqdm",
    "requests",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["util*"]